+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 * [FEATURE] Various snackbar undo actions #62
 * [FIX] prompt user to save unsaved changes in text file #159
 * [FIX] Not visible if a resource is a user or group in share dropdown #168
-
+* [FIX] Share link from search context, no ui response #144
 
 ## 3.1.2
 **Maintainer**: balloon-team <opensource@gyselroth.net>\